requirements.txt
pillow>=6.1
pytest>=5
python-telegram-bot[socks]
python-telegram-bot>=12
psycopg2-binary>=2.8.3
sqlalchemy>=1.3
sqlalchemy-utils>=0.34.2
